# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Reset Server",
        "description": "RestServer Node.js API",
        "license": "ISC"
    },
    "servers": [
        {
            "url": "https://rest-server-v1-api.herokuapp.com",
            "description": "Entorno de producción"
        }
    ],
    "basePath": "",
...
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/api/users": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Crear nuevo usuario en el servidor",
                "description": "Crear nuevo usuario en el servidor",
                "operationId": "userPost",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            },
                            "example": {
                                "name": "Test",
                                "email": "test@correo.com",
                                "password": "123456",
                                "role": "USER_ROLE"
                            }
                        }
                    }
                },
                "application/xml": {
                    "schema": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario creado correctamente!"
                    }
                }
            },
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Buscar usuarios en el servidor",
                "description": "Buscar usuarios en el servidor",
                "operationId": "userGet",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Indique el numero desde el que quiere partir",
                        "required": false,
                        "type": "number",
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Indique la cantidad de elementos",
                        "required": false,
                        "type": "number",
                        "collectionFormat": "multi"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario creado correctamente!",
                        "examples": {
                            "application/json": {
                                "total": 20,
                                "userGet": "array"
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Crear nuevo usuario en el servidor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "Eliminar usuario por id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado",
                        "examples": {
                            "application/json": {
                                "user": "array"
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Actualizar usuario",
                "description": "Actualizar usuario",
                "operationId": "userPut",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "Pasos para crear un usuario"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            },
                            "example": {
                                "name": "Test",
                                "email": "test@correo.com",
                                "password": "123456",
                                "role": "USER_ROLE"
                            }
                        }
                    }
                },
                "application/xml": {
                    "schema": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario Actualizado",
                        "examples": {
                            "application/json": {
                                "userPut": "array"
                            }
                        }
                    },
                    "400": {
                        "description": "Id invalido"
                    },
                    "404": {
                        "description": "Id no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "x-token",
                "scheme": "bearer",
                "in": "path",
                "bearerFormat": "JWT"
            }
        },
...
        "schemas": {
            "User": {
                "type": "object",
                "required": [
                    "name",
                    "email",
                    "password",
                    "role"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "uniqueItems": false,
                        "example": "Test"
                    },
                    "email": {
                        "type": "string",
                        "uniqueItems": true,
                        "example": "test@correo.com"
                    },
                    "password": {
                        "type": "string",
                        "uniqueItems": false,
                        "example": "123456"
                    },
                    "role": {
                        "type": "string",
                        "uniqueItems": false,
                        "example": "USER_ROLE"
                    },
                    "state": {
                        "type": "boolean",
                        "uniqueItems": false
                    },
                    "google": {
                        "type": "boolean",
                        "uniqueItems": false
                    }
                }
            },
...
        }
    },
...
}